+++ src/neo4j_graphrag/experimental/components/data_loader.py
@@ -12,14 +12,14 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-"""Document loaders: base class, PDF, Markdown, LiteParse, and extension-based dispatch."""
+"""Document loaders: base class, PDF, and Markdown."""
 
 from __future__ import annotations
 
 import io
 from abc import abstractmethod
 from pathlib import Path
-from typing import Any, Dict, Optional, Union, cast
+from typing import Dict, Optional, Union, cast
 
 import fsspec
 import pypdf
@@ -27,14 +27,6 @@ from fsspec import AbstractFileSystem
 from fsspec.implementations.local import LocalFileSystem
 
 from neo4j_graphrag.exceptions import MarkdownLoadError, PdfLoaderError
-
-try:
-    from liteparse import LiteParse as _LiteParse  # type: ignore[import]
-
-    _LITEPARSE_AVAILABLE = True
-except ImportError:
-    _LiteParse = None  # type: ignore[assignment,misc]
-    _LITEPARSE_AVAILABLE = False
 from neo4j_graphrag.experimental.components.types import (
     DocumentInfo,
     DocumentType,
@@ -143,93 +135,5 @@ class MarkdownLoader(DataLoader):
                 path=filepath,
                 metadata=self.get_document_metadata(text, metadata),
                 document_type=DocumentType.MARKDOWN,
-            ),
-        )
-
-
-class LiteParseLoader(DataLoader):
-    """Loads and parses PDF files using LiteParse (local, no cloud dependency).
-
-    LiteParse uses a compiled Rust core and optional Tesseract OCR to extract text
-    from PDFs, DOCX, XLSX, PPTX, and images.  It runs fully offline — no API key
-    or network access required.
-
-    Requires the ``liteparse`` optional extra::
-
-        pip install "neo4j-graphrag[liteparse]"
-
-    Args:
-        ocr_enabled: Enable Tesseract OCR for scanned pages.
-        ocr_server_url: URL of a remote OCR server (optional).
-        ocr_language: Tesseract language code, e.g. ``"eng"``, ``"fra"``.
-        dpi: Rendering resolution for OCR (default 300).
-        target_pages: Page range string, e.g. ``"1-5,10,15-20"``.
-        password: Password for encrypted PDFs.
-    """
-
-    def __init__(
-        self,
-        ocr_enabled: bool = False,
-        ocr_server_url: Optional[str] = None,
-        ocr_language: Optional[str] = None,
-        dpi: Optional[int] = None,
-        target_pages: Optional[str] = None,
-        password: Optional[str] = None,
-    ) -> None:
-        optional = {
-            "ocr_server_url": ocr_server_url,
-            "ocr_language": ocr_language,
-            "dpi": dpi,
-            "target_pages": target_pages,
-            "password": password,
-        }
-        self._kwargs: Dict[str, Any] = {
-            "ocr_enabled": ocr_enabled,
-            **{k: v for k, v in optional.items() if v is not None},
-        }
-        self._parser: Any = None  # lazily initialised
-
-    def _get_parser(self) -> Any:
-        if not _LITEPARSE_AVAILABLE:
-            raise ImportError(
-                "liteparse is required for LiteParseLoader. "
-                'Install it with: pip install "neo4j-graphrag[liteparse]"'
-            )
-        if self._parser is None:
-            self._parser = _LiteParse(**self._kwargs)
-        return self._parser
-
-    def load_file(self, filepath: str, fs: AbstractFileSystem) -> str:
-        """Parse a document and return the full extracted text."""
-        parser = self._get_parser()  # ImportError propagates; not caught below
-        try:
-            if is_default_fs(fs):
-                result = parser.parse(filepath)
-            else:
-                with fs.open(filepath, "rb") as fp:
-                    result = parser.parse(fp.read())
-            return str(result.text)
-        except Exception as e:
-            raise PdfLoaderError(e) from e
-
-    async def run(
-        self,
-        filepath: Union[str, Path],
-        metadata: Optional[Dict[str, str]] = None,
-        fs: Optional[Union[AbstractFileSystem, str]] = None,
-    ) -> LoadedDocument:
-        if not isinstance(filepath, str):
-            filepath = str(filepath)
-        if isinstance(fs, str):
-            fs = fsspec.filesystem(fs)
-        elif fs is None:
-            fs = LocalFileSystem()
-        text = self.load_file(filepath, fs)
-        return LoadedDocument(
-            text=text,
-            document_info=DocumentInfo(
-                path=filepath,
-                metadata=self.get_document_metadata(text, metadata),
-                document_type=DocumentType.PDF,
             ),
         )
